feat: migrate dust app internals to vaults
