+++ front/components/vaults/VaultCreateAppModal.tsx
@@ -5,21 +5,23 @@ import {
   Page,
   TextArea,
 } from "@dust-tt/sparkle";
-import type { APIError, WorkspaceType } from "@dust-tt/types";
+import type { APIError, VaultType, WorkspaceType } from "@dust-tt/types";
 import { useRouter } from "next/router";
 import { useContext, useState } from "react";
 
 import { SendNotificationsContext } from "@app/components/sparkle/Notification";
 import { useApps } from "@app/lib/swr/apps";
 import { MODELS_STRING_MAX_LENGTH } from "@app/lib/utils";
-import type { PostAppResponseBody } from "@app/pages/api/w/[wId]/apps";
+import type { PostAppResponseBody } from "@app/pages/api/w/[wId]/vaults/[vId]/apps";
 
 export const VaultCreateAppModal = ({
   owner,
+  vault,
   isOpen,
   setIsOpen,
 }: {
   owner: WorkspaceType;
+  vault: VaultType;
   isOpen: boolean;
   setIsOpen: (isOpen: boolean) => void;
 }) => {
@@ -37,7 +39,7 @@ export const VaultCreateAppModal = ({
     description: null,
   });
 
-  const { apps, mutateApps } = useApps(owner);
+  const { apps, mutateApps } = useApps({ owner, vault });
 
   const onSave = async () => {
     let nameError: string | null = null;
@@ -63,7 +65,7 @@ export const VaultCreateAppModal = ({
     });
 
     if (name && description && !nameError && !descriptionError) {
-      const res = await fetch(`/api/w/${owner.sId}/apps`, {
+      const res = await fetch(`/api/w/${owner.sId}/vaults/${vault.sId}/apps`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
