+++ front/components/vaults/VaultSideBarMenu.tsx
@@ -554,7 +554,11 @@ const VaultAppSubMenu = ({
 
   const categoryDetails = DATA_SOURCE_OR_VIEW_SUB_ITEMS[category];
 
-  const { isAppsLoading, apps } = useApps(owner, !isExpanded);
+  const { isAppsLoading, apps } = useApps({
+    owner,
+    disabled: !isExpanded,
+    vault,
+  });
 
   return (
     <Tree.Item
