+++ front/pages/api/w/[wId]/vaults/[vId]/apps/[aId]/runs/[runId]/blocks/[type]/[name]/index.ts
@@ -1,30 +1,50 @@
-import type { RunType, WithAPIErrorResponse } from "@dust-tt/types";
+import type { BlockType, RunType, WithAPIErrorResponse } from "@dust-tt/types";
 import { CoreAPI } from "@dust-tt/types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import config from "@app/lib/api/config";
+import apiConfig from "@app/lib/api/config";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { AppResource } from "@app/lib/resources/app_resource";
 import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
 
-export type GetRunStatusResponseBody = {
+export const config = {
+  api: {
+    responseLimit: "8mb",
+  },
+};
+
+export type GetRunBlockResponseBody = {
   run: RunType | null;
 };
 
 async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<GetRunStatusResponseBody>>,
+  res: NextApiResponse<WithAPIErrorResponse<GetRunBlockResponseBody>>,
   auth: Authenticator
-) {
+): Promise<void> {
+  const vaultId = req.query.vId as string;
+
   const app = await AppResource.fetchById(auth, req.query.aId as string);
-  if (!app) {
+
+  if (!app || app.vault.sId !== vaultId) {
     return apiError(req, res, {
       status_code: 404,
       api_error: {
         type: "app_not_found",
         message: "The app was not found.",
+      },
+    });
+  }
+
+  if (!app.canWrite(auth)) {
+    return apiError(req, res, {
+      status_code: 403,
+      api_error: {
+        type: "app_auth_error",
+        message:
+          "Retrieving content of runs requires write access to the app's vault.",
       },
     });
   }
@@ -42,28 +62,20 @@ async function handler(
         res.status(200).json({ run: null });
         return;
       }
-
-      const coreAPI = new CoreAPI(config.getCoreAPIConfig(), logger);
-      const run = await coreAPI.getRunStatus({
+      const coreAPI = new CoreAPI(apiConfig.getCoreAPIConfig(), logger);
+      const run = await coreAPI.getRunBlock({
         projectId: app.dustAPIProjectId,
         runId: runId,
+        blockType: req.query.type as BlockType,
+        blockName: req.query.name as string,
       });
-      if (run.isErr()) {
-        if (run.error.code === "run_not_found") {
-          return apiError(req, res, {
-            status_code: 404,
-            api_error: {
-              type: "run_not_found",
-              message: "The run was not found.",
-            },
-          });
-        }
 
+      if (run.isErr()) {
         return apiError(req, res, {
           status_code: 500,
           api_error: {
             type: "internal_server_error",
-            message: "The run status retrieval failed.",
+            message: "The run block retrieval failed.",
             app_error: run.error,
           },
         });
