+++ front/pages/api/w/[wId]/vaults/[vId]/apps/[aId]/runs/[runId]/status.ts
@@ -1,50 +1,42 @@
-import type { BlockType, RunType, WithAPIErrorResponse } from "@dust-tt/types";
+import type { RunType, WithAPIErrorResponse } from "@dust-tt/types";
 import { CoreAPI } from "@dust-tt/types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import apiConfig from "@app/lib/api/config";
+import config from "@app/lib/api/config";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { AppResource } from "@app/lib/resources/app_resource";
 import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
 
-export const config = {
-  api: {
-    responseLimit: "8mb",
-  },
-};
-
-export type GetRunBlockResponseBody = {
+export type GetRunStatusResponseBody = {
   run: RunType | null;
 };
 
 async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<GetRunBlockResponseBody>>,
+  res: NextApiResponse<WithAPIErrorResponse<GetRunStatusResponseBody>>,
   auth: Authenticator
-): Promise<void> {
-  // Only the users that are `builders` for the current workspace can retrieve the content of runs
-  // block outputs.
-  if (!auth.isBuilder()) {
-    return apiError(req, res, {
-      status_code: 403,
-      api_error: {
-        type: "app_auth_error",
-        message:
-          "Only the users that are `builders` for the current workspace can introspect runs.",
-      },
-    });
-  }
+) {
+  const vaultId = req.query.vId as string;
 
   const app = await AppResource.fetchById(auth, req.query.aId as string);
-
-  if (!app) {
+  if (!app || app.vault.sId !== vaultId) {
     return apiError(req, res, {
       status_code: 404,
       api_error: {
         type: "app_not_found",
         message: "The app was not found.",
+      },
+    });
+  }
+
+  if (!app.canRead(auth)) {
+    return apiError(req, res, {
+      status_code: 403,
+      api_error: {
+        type: "app_auth_error",
+        message: "Reading the app requires read access to the app's vault.",
       },
     });
   }
@@ -62,20 +54,28 @@ async function handler(
         res.status(200).json({ run: null });
         return;
       }
-      const coreAPI = new CoreAPI(apiConfig.getCoreAPIConfig(), logger);
-      const run = await coreAPI.getRunBlock({
+
+      const coreAPI = new CoreAPI(config.getCoreAPIConfig(), logger);
+      const run = await coreAPI.getRunStatus({
         projectId: app.dustAPIProjectId,
         runId: runId,
-        blockType: req.query.type as BlockType,
-        blockName: req.query.name as string,
       });
-
       if (run.isErr()) {
+        if (run.error.code === "run_not_found") {
+          return apiError(req, res, {
+            status_code: 404,
+            api_error: {
+              type: "run_not_found",
+              message: "The run was not found.",
+            },
+          });
+        }
+
         return apiError(req, res, {
           status_code: 500,
           api_error: {
             type: "internal_server_error",
-            message: "The run block retrieval failed.",
+            message: "The run status retrieval failed.",
             app_error: run.error,
           },
         });
