+++ front/pages/w/[wId]/vaults/[vaultId]/apps/[aId]/datasets/new.tsx
@@ -114,7 +114,7 @@ export default function NewDatasetView({
 
   const handleSubmit = async () => {
     setLoading(true);
-    const res = await fetch(`/api/w/${owner.sId}/apps/${app.sId}/datasets`, {
+    const res = await fetch(`/api/w/${owner.sId}/vaults/${app.vault.sId}/apps/${app.sId}/datasets`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
